changed cudart typing to use size_t

+++ gpu/gpu_info_cuda.h
@@ -33,9 +33,9 @@ typedef struct nvmlMemory_st {
 } nvmlMemory_t;
 
 typedef struct cudartMemory_st {
-  unsigned long long total;
-  unsigned long long free;
-  unsigned long long used;
+  size_t total;
+  size_t free;
+  size_t used;
 } cudartMemory_t;
 
 typedef struct cudartDriverVersion {
